break out related terms component
